requirements.txt
# Core ML and Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.31.0
tensorflow>=2.13.0

# Technical Analysis
ta>=0.10.0
# talib-binary>=0.4.24  # Commented out - requires manual installation
mplfinance>=0.12.10b0

# Data Sources and APIs
alpaca-py>=0.13.0
polygon-api-client>=1.12.0
yfinance>=0.2.0
ccxt>=4.0.0

# Real-time Processing
asyncio-mqtt>=0.13.0
websockets>=11.0.0
aiohttp>=3.8.0
fastapi>=0.101.0
uvicorn>=0.23.0

# Database and Caching
psycopg2-binary>=2.9.0
redis>=4.6.0
sqlalchemy>=2.0.0
alembic>=1.11.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0
mlflow>=2.6.0
wandb>=0.15.0

# Backtesting and Analysis
backtrader>=1.9.78
vectorbt>=0.25.0
empyrical>=0.5.0
# pyfolio>=0.9.2  # Commented out - may have compatibility issues

# Configuration and Utils
pydantic>=2.1.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter for Analysis
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Production
gunicorn>=21.2.0
celery>=5.3.0
flower>=2.0.0 